requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
pandas
numpy
scikit-learn
pymongo
gunicorn